complete engine discovery for docker tests
If a test method or a class is annotated with @Dockerized, then they
will be discovered by the docker-test-engine. To disable dockerized
test methods in presence of the default engine, you can use EnabledIf
 with a variant like EnvironmentVariable set. For this last case, in
the docker container, the env variable is set and only the new test
engine is enabled so that it runs only the discovered methods.