execute method is now implemented
the docker container is started once with gradle --deamon so that only
the daemon is started, but the container remains up. Then it is made
sure that all classes are compiled so that the classpath is correctly
added when calling the console launcher. After parsing the console result
the outcome is definitely not the one JUnit5 gives you. this is because
the you should have to reproduce all the discovery mechanism of JUnit5.
Until now, every method is called separately, aggregating results from
parametric ones.