modified engine to inherit from HierarchicalTestEngine
HierarchicalTestEngine is a class made available from the junit-jupiter
engine and it is very handy to avoid the execution phase. Every descriptor
must implement the Node interface and must override the execute method
if you want it to be executed in the execution phase.
A limit I encountered is that you cannot easily change the TestExecutionResult
for a container type, even if it is usually marked as successful.