dockerized annotation enriched with container name and image
now inside the dockerized annotation you can specify the container
name and image of the created container. Every descriptor can access
this information thanks to the DockerizableDescriptor interface. at
the start the root descriptor finds all the container and create
a map (name, image). After that, when containers are actually started,
the container ids are stored in a map (name, id) used by method
descriptors when running test.