testcontainer results do not depend on test results
now testcontainer results are all successful as they do not
depend on the results of the contained tests. this should be the
standard, since a container failure means that some initialization
throws an exception or something like that, but not that any of
its children are failed. In this way you can distinguish a failure
initialization from contained tests failure