remove environment variable constraint for docker image
the environment variable DOCKER_ONLY is not necessary anymore because
a conditional test now depends on the system property
testingEnvironment, that must be "docker" in a dockerized env.
the system property can be easily set with -D both in gradle
and in java. because tests in gradle could be executed in a different
jvm, you need this workaround if you want to execute them in the
host machine:

https://stackoverflow.com/questions/21406265/how-to-give-system-property-to-my-test-via-gradle-and-d

already converted to a groovy method in build.gradle (and commented)